+++ app/(gcforms)/[locale]/(form administration)/form-builder/actions.ts
@@ -20,6 +20,7 @@ import {
   updateSecurityAttribute,
   updateResponseDeliveryOption,
   updateFormPurpose,
+  updateFormSaveAndResume,
 } from "@lib/templates";
 import { serverTranslation } from "@i18n";
 import { revalidatePath } from "next/cache";
@@ -33,6 +34,7 @@ export type CreateOrUpdateTemplateType = {
   deliveryOption?: DeliveryOption;
   securityAttribute?: SecurityAttribute;
   formPurpose?: FormPurpose;
+  saveAndResume?: boolean;
 };
 
 // Public facing functions - they can be used by anyone who finds the associated server action identifer
@@ -49,7 +51,7 @@ export const createOrUpdateTemplate = AuthenticatedAction(
       formPurpose,
     }: CreateOrUpdateTemplateType
   ): Promise<{
-    formRecord: FormRecord | null;
+    formRecord: { id: string; updatedAt: string | undefined } | null;
     error?: string;
   }> => {
     try {
@@ -66,7 +68,7 @@ export const createOrUpdateTemplate = AuthenticatedAction(
         });
       }
 
-      const response = await createDbTemplate({
+      const formRecord = await createDbTemplate({
         userID: session.user.id,
         formConfig: formConfig,
         name: name,
@@ -75,15 +77,13 @@ export const createOrUpdateTemplate = AuthenticatedAction(
         formPurpose: formPurpose,
       });
 
-      if (!response) {
-        throw new Error(
-          `Template API response was null. Request information: { formConfig: ${formConfig}, name: ${name}, deliveryOption: ${deliveryOption}, securityAttribute: ${securityAttribute}`
-        );
+      if (!formRecord) {
+        throw new Error("Failed to create template");
       }
 
-      return { formRecord: response };
-    } catch (e) {
-      return { formRecord: null, error: (e as Error).message };
+      return { formRecord: { id: formRecord.id, updatedAt: formRecord.updatedAt } };
+    } catch (_) {
+      return { formRecord: null, error: "error" };
     }
   }
 );
@@ -107,11 +107,11 @@ export const updateTemplate = AuthenticatedAction(
       formPurpose?: FormPurpose;
     }
   ): Promise<{
-    formRecord: FormRecord | null;
+    formRecord: { id: string; updatedAt: string | undefined } | null;
     error?: string;
   }> => {
     try {
-      const response = await updateDbTemplate({
+      const formRecord = await updateDbTemplate({
         formID: formID,
         formConfig: formConfig,
         name: name,
@@ -119,14 +119,14 @@ export const updateTemplate = AuthenticatedAction(
         securityAttribute: securityAttribute,
         formPurpose: formPurpose,
       });
-      if (!response) {
-        throw new Error(
-          `Template API response was null. Request information: { formConfig: ${formConfig}, name: ${name}, deliveryOption: ${deliveryOption}, securityAttribute: ${securityAttribute}`
-        );
+
+      if (!formRecord) {
+        throw new Error("Failed to update template");
       }
-      return { formRecord: response };
-    } catch (error) {
-      return { formRecord: null, error: (error as Error).message };
+
+      return { formRecord: { id: formRecord.id, updatedAt: formRecord.updatedAt } };
+    } catch (_) {
+      return { formRecord: null, error: "error" };
     }
   }
 );
@@ -193,6 +193,35 @@ export const updateTemplateFormPurpose = AuthenticatedAction(
       if (!response) {
         throw new Error(
           `Template API response was null. Request information: { ${formID}, ${formPurpose} }`
+        );
+      }
+
+      return { formRecord: response };
+    } catch (error) {
+      return { formRecord: null, error: (error as Error).message };
+    }
+  }
+);
+
+export const updateTemplateFormSaveAndResume = AuthenticatedAction(
+  async (
+    _,
+    {
+      id: formID,
+      saveAndResume,
+    }: {
+      id: string;
+      saveAndResume: boolean;
+    }
+  ): Promise<{
+    formRecord: FormRecord | null;
+    error?: string;
+  }> => {
+    try {
+      const response = await updateFormSaveAndResume(formID, saveAndResume);
+      if (!response) {
+        throw new Error(
+          `Template API response was null. Request information: { ${formID}, ${saveAndResume} }`
         );
       }