+++ prisma/migrations/20250130165334_create_template_user_view/migration.sql
@@ -1,0 +1,3 @@
+CREATE VIEW "TemplateUserView" AS
+SELECT "A" as templateId, "B" as userId
+FROM "_TemplateToUser";